# packages/api/openapi.yaml
openapi: 3.1.0
info:
  title: Teaching App API
  version: 1.0.0
  description: REST API for the Teaching App - An async learning management platform
servers:
  - url: /api
    description: Default server
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: accessToken
      description: HttpOnly access token cookie set on login
  schemas:
...
    LessonSummary:
      type: object
      required: [type, lessonId, summary, sources]
      properties:
        type: { type: string, enum: [summary] }
        lessonId: { type: string, format: uuid }
        summary: { type: string }
        sources:
          type: array
          items:
            type: object
            required: [type, id, title]
            properties:
              type: { type: string, enum: [lesson, resource] }
              id: { type: string, format: uuid }
              title: { type: string }
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: Invalid credentials
...
paths:
...
  /modules/{moduleId}/lessons/{lessonId}/summary:
    post:
      tags: [Lessons]
      summary: Generate a lesson study summary from lesson text and all lesson PDFs
      operationId: generateLessonSummary
      security: [{ cookieAuth: [] }]
      parameters:
        - {
            name: moduleId,
            in: path,
            required: true,
            schema: { type: string, format: uuid },
          }
        - {
            name: lessonId,
            in: path,
            required: true,
            schema: { type: string, format: uuid },
          }
      responses:
        "200":
          {
            description: Generated Markdown summary,
            content:
              {
                application/json:
                  {
                    schema:
                      {
                        type: object,
                        properties:
                          {
                            data:
                              { $ref: "#/components/schemas/LessonSummary" },
                          },
                      },
                  },
              },
          }
        "401":
          {
            description: Unauthorized,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/ErrorResponse" } },
              },
          }
        "403":
          {
            description: Forbidden,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/ErrorResponse" } },
              },
          }
        "422":
          {
            description: No material or supported input bound exceeded,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/ErrorResponse" } },
              },
          }
        "502":
          {
            description: AI provider failure,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/ErrorResponse" } },
              },
          }